Updating sorting and episode files
