+++ app/controllers/podcasts/index.js
@@ -1,5 +1,10 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-  sortProperties: ['title']
+  sortedPodcasts: function() {
+    return Ember.ArrayProxy.extend(Ember.SortableMixin).create({
+      sortProperties: ['title'],
+      content: this.get('model')
+    });
+  }.property('model')
 });
