+++ app/templates/podcasts/show.hbs
@@ -20,4 +20,4 @@
   <!-- <p>{{model.description}}</p> -->
 </div>
 
-{{render "episodes" model.episodes}}
+{{render "episodes" controller.sortedEpisodes}}
